requirements.txt
# Flask Core
Flask==3.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication
Flask-Login==0.6.3
bcrypt==4.1.2

# Forms & Validation
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0

# Email
Flask-Mail==0.9.1

# Environment
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2
